Operaciones CRUD con respecto a los artículos

+++ html/webservices/acciones/adminarticulo.accion.php
@@ -95,54 +95,72 @@
                     $post_contenido,
                     $post_estado,
                     $post_idusuario,
-                    $post_tags,
-                    $post_categorias,
-                    $post_categorias_eliminadas
+                    $post_tags
                 )) {
-                    $strSql = "";
-                    $parametros = null;
-                    if (trim($post_contenido) === "") {
-                        $strSql = "
-                            UPDATE mod_articulos
-                            SET titulo = ?,
-                                fecha = now(),
-                                estado = ?,
-                                idusuario = ?,
-                                tags = ?
-                            WHERE idarticulo = ?
-                        "; 
-                        $parametros = [
-                            $post_titulo,
-                            $post_estado,
-                            $post_idusuario,
-                            $post_tags,
-                        ]
+
+                    // Verificamos que el artículo existe
+                    $queryExisteArticulo = "select ifnull(count(*), 0) as existeArticulo
+                                            from mod_articulos
+                                            where idarticulo = ?;";
+
+                    $existeArticulo = $f->getQueryData($queryExisteArticulo, [$post_idarticulo]);
+
+                    if ($existeArticulo["data"][0]["existeArticulo"]) {
+
+                        $strSql = "";
+                        $parametros = null;
+                        if (trim($post_contenido) === "") {
+                            $strSql = "
+                                UPDATE mod_articulos
+                                SET titulo = ?,
+                                    fecha = now(),
+                                    estado = ?,
+                                    idusuario = ?,
+                                    tags = ?
+                                WHERE idarticulo = ?
+                            "; 
+                            $parametros = [
+                                $post_titulo,
+                                $post_estado,
+                                $post_idusuario,
+                                $post_tags,
+                                $post_idarticulo
+                            ];
+                        } else {
+                            $strSql = "
+                                UPDATE mod_articulos
+                                SET titulo = ?,
+                                    contenido = ?,
+                                    fecha = now(),
+                                    estado = ?,
+                                    idusuario = ?,
+                                    tags = ?
+                                WHERE idarticulo = ?
+                            "; 
+                            $parametros = [
+                                $post_titulo,
+                                $post_contenido,
+                                $post_estado,
+                                $post_idusuario,
+                                $post_tags,
+                                $post_idarticulo
+                            ];
+                        }
+                        // cambiar los datos del articulo 
+                        $queryData = $f->exeQuery($strSql, $parametros);
+                        if (!$queryData["error"]) {
+                            $text = "Artículo editado con éxito.";
+                            $type = "success";
+                            $title = "Éxito";
+                            $datareturn = $parametros;
+                        }
+
                     } else {
-                        $strSql = "
-                            UPDATE mod_articulos
-                            SET titulo = ?,
-                                contenido = ?,
-                                fecha = now(),
-                                estado = ?,
-                                idusuario = ?,
-                                tags = ?
-                            WHERE idarticulo = ?
-                        "; 
-                        $parametros = [
-                            $post_titulo,
-                            $post_contenido,
-                            $post_estado,
-                            $post_idusuario,
-                            $post_tags,
-                        ]
-                    }
-
-                    // cambiar los datos del articulo 
-                    $queryData = $f->getQueryData();
-                    if (!$queryData["error"]) {
-                        // cambiar las referencias de categorias pertenecientes al articulo
-
-                    }
+                        $text = "Al parecer el artículo que desea editar no existe. Intenta con otro.";
+                        $type = "error";
+                        $title = "Éxito";
+                        $datareturn = $post_idarticulo;
+                    }                    
 
                 } else {
                     $text = "No envio todos los datos necesarios para crear el articulo. Intentelo de nuevo";
@@ -150,20 +168,35 @@
                 break;
             case "eliminar":
                 if (isset($post_idarticulo)) {
-                    $queryEliminarArticulo = "DELETE FROM mod_articulos WHERE idarticulo = ?";
-                    $parametros = [$post_idarticulo];
-                    $queryData = $f->exeQuery($queryEliminarArticulo, $parametros);
 
-                    if (!$queryData["error"]) {
-                        $text = "Artículo eliminado con éxito.";
-                        $type = "success";
-                        $title = "Éxito";
-                        $datareturn = $parametros;
+                    $queryExisteArticulo = "select ifnull(count(*), 0) as existeArticulo
+                                            from mod_articulos
+                                            where idarticulo = ?;";
+
+                    $existeArticulo = $f->getQueryData($queryExisteArticulo, [$post_idarticulo]);
+
+                    if ($existeArticulo["data"][0]["existeArticulo"]) {
+                        // verificamos si el artículo existe y si es así eliminamos el artículo.
+                        $queryEliminarArticulo = "DELETE FROM mod_articulos WHERE idarticulo = ?";
+                        $parametros = [$post_idarticulo];
+                        $queryData = $f->exeQuery($queryEliminarArticulo, $parametros);
+
+                        if (!$queryData["error"]) {
+                            $text = "Artículo eliminado con éxito.";
+                            $type = "success";
+                            $title = "Éxito";
+                            $datareturn = $parametros;
+                        } else {
+                            $text = "No se ha podido eliminar el artículo: " + $queryData['error'];
+                            $type = "error";
+                            $title = "Éxito";
+                            $datareturn = $parametros;
+                        }
                     } else {
-                        $text = "No se ha podido eliminar el artículo: " + $queryData['error'];
+                        $text = "Al parecer el artículo que desea eliminar no existe. Intenta con otro.";
                         $type = "error";
-                        $title = "Éxito";
-                        $datareturn = $parametros;
+                        $title = "Error";
+                        $datareturn = $post_idarticulo;
                     }
                 } else {
                     $text = "No envio todos los datos necesarios para crear el articulo. Intentelo de nuevo";
@@ -184,10 +217,5 @@
         $corchetes = ["[", "]", " ", "\""];
         $x = str_replace($corchetes, "", $c);
         return $x;
-    }
-
-    // algoritmo para saber cual categoria fue eliminada 
-    function getEliminatedCategory($idArticulo, ) {
-
     }
 ?>
